+++ lib/Params.mli
@@ -84,16 +84,18 @@ type elements_collection_record_pat = {wildcard: Fmt.t}
 val get_record_expr :
   Conf.t -> elements_collection * elements_collection_record_expr
 
-val get_list_expr : Conf.t -> elements_collection
+val get_list_expr : Conf.t -> box_collec:Fmt.box -> elements_collection
 
-val get_array_expr : Conf.t -> elements_collection
+val get_array_expr : Conf.t -> box_collec:Fmt.box -> elements_collection
 
 val get_record_pat :
   Conf.t -> ctx:Ast.t -> elements_collection * elements_collection_record_pat
 
-val get_list_pat : Conf.t -> ctx:Ast.t -> elements_collection
+val get_list_pat :
+  Conf.t -> ctx:Ast.t -> box_collec:Fmt.box -> elements_collection
 
-val get_array_pat : Conf.t -> ctx:Ast.t -> elements_collection
+val get_array_pat :
+  Conf.t -> ctx:Ast.t -> box_collec:Fmt.box -> elements_collection
 
 type if_then_else =
   { box_branch: Fmt.t -> Fmt.t